% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_candidate_info}
\alias{all_candidate_info}
\title{Data from MA OCPF about candidates including name, ID, office}
\format{
tibble data.frame with 4143 rows and 34 columns
}
\usage{
all_candidate_info
}
\description{
Data set from running the get_local_candidates() function
}
\keyword{datasets}
