% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_framingham_candidates}
\alias{all_framingham_candidates}
\title{Data from MA OCPF about Framingham candidates including name, ID, office}
\format{
tibble data.frame with 72 rows and 34 columns
}
\usage{
all_framingham_candidates
}
\description{
Data set from running the get_local_candidates() function and filtering for Candidate_City == "Framingham"
}
\keyword{datasets}
