% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candidate_contribution_data.R
\name{get_latest_candidate_contributions}
\alias{get_latest_candidate_contributions}
\title{Get latest candidate contributions by candidate ID}
\usage{
get_latest_candidate_contributions(id, start_date = NULL)
}
\arguments{
\item{id}{character string with candidate MA OCPF ID}

\item{start_date}{string in yyyy-mm-dd format or date object for contributions on or after a certain date, defaults to NULL (all candidate data downloads; this function filters it)}
}
\value{
data frame with contribution information
}
\description{
Get latest candidate contributions by candidate ID
}
