% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_candidates.R
\name{get_local_candidates}
\alias{get_local_candidates}
\title{Get all candidates throughout Massachusetts - downloads full data each time
Suggest running once per session with defaults and then filtering later on community or office. Data downloads to temp subdirectory of your working directory (will be created if it doesn't exist).}
\usage{
get_local_candidates(
  place = "",
  office = "",
 
    the_url = "https://ocpf2.blob.core.windows.net/downloads/data/registered-candidates.zip"
)
}
\arguments{
\item{place}{character string defaults to "" to load all}

\item{office}{character string defaults to "" to load all}

\item{the_url}{character string for OCPF url to download zip file}
}
\value{

}
\description{
Get all candidates throughout Massachusetts - downloads full data each time
Suggest running once per session with defaults and then filtering later on community or office. Data downloads to temp subdirectory of your working directory (will be created if it doesn't exist).
}
